updated jelix to 1.2b1pre

+++ lib/jelix/plugins/tpl/html/function.formdatafull.php
@@ -3,9 +3,9 @@
 * @package      jelix
 * @subpackage   jtpl_plugin
 * @author       Laurent Jouanneau
-* @contributor  Dominique Papin, Julien Issler
-* @copyright    2007-2008 Laurent Jouanneau, 2007 Dominique Papin
-* @copyright    2008 Julien Issler
+* @contributor  Dominique Papin, Julien Issler, Brunto
+* @copyright    2007-2010 Laurent Jouanneau, 2007 Dominique Papin
+* @copyright    2008 Julien Issler, 2010 Brunto
 * @link         http://www.jelix.org
 * @licence      GNU Lesser General Public Licence see LICENCE file or http://www.gnu.org/licenses/lgpl.html
 */
@@ -21,22 +21,23 @@ function jtpl_function_html_formdatafull($tpl, $form)
     echo '<table class="jforms-table" border="0">';
 
     foreach( $form->getControls() as $ctrlref=>$ctrl){
-        if($ctrl->type == 'submit' || $ctrl->type == 'reset' || $ctrl->type == 'hidden' || $ctrl->type == 'captcha') continue;
+        if($ctrl->type == 'submit' || $ctrl->type == 'reset' || $ctrl->type == 'hidden' || $ctrl->type == 'captcha' || $ctrl->type == 'secretconfirm') continue;
         if(!$form->isActivated($ctrlref)) continue;
 
         echo '<tr><th scope="row">';
         echo htmlspecialchars($ctrl->label);
         echo '</th><td>';
         $value = $ctrl->getDisplayValue($form->getData($ctrlref));
-        if(is_array($value)){
+        if (is_array($value)) {
             $s ='';
-            foreach($value as $v){
+            foreach ($value as $v) {
                 $s.=','.htmlspecialchars($v);
             }
             echo substr($s, 1);
-        }elseif($ctrl->datatype instanceof jDatatypeHtml)
+        }
+        else if ($ctrl->isHtmlContent())
             echo $value;
-        else if($ctrl->type == 'textarea')
+        else if ($ctrl->type == 'textarea')
             echo nl2br(htmlspecialchars($value));
         else
             echo htmlspecialchars($value);
